# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Delivery SaaS API
  version: 1.0.0
  description: Contrato principal multi-tenant. Valores monetarios de dominio usam centavos.
servers:
  - url: /
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: delivery_session
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
...
paths:
...
  /api/customer/stores/{slug}/auth/register:
    post:
      summary: Cria conta no tenant e emite cookies de sessao e CSRF
      responses: { '201': { description: Perfil publico da conta }, '409': { description: Conta ja existente } }
  /api/customer/stores/{slug}/auth/login:
    post:
      summary: Autentica telefone e senha no tenant
      responses: { '200': { description: Perfil e cookies emitidos }, '401': { description: Credenciais invalidas } }
...
  /api/customer/stores/{slug}/auth/logout:
    post:
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Revoga sessao e limpa cookies
      responses: { '200': { description: Sessao encerrada } }
  /api/customer/stores/{slug}/auth/profile:
    put:
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Atualiza nome, email, nascimento e genero
      responses: { '200': { description: Perfil atualizado } }
  /api/customer/stores/{slug}/auth/password:
    put:
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Troca senha e revoga sessoes
      responses: { '200': { description: Nova autenticacao obrigatoria } }
  /api/customer/stores/{slug}/auth/password/request:
    post:
      summary: Solicita recuperacao sem revelar existencia da conta
      responses: { '202': { description: Solicitacao aceita }, '503': { description: Provider OTP indisponivel } }
...